Added සිංහල, Sinhala language project to Readme
I'm a developer from Sri Lanka and I thought of adding the Sinhala language support to the Redux documentation. Currently Redux is becoming popular within the country and I think it would be useful for them to understand it from their native language. Please be kind enough to add this link so that they can easily find it and I will be finished with the translations in near future. 
Regards 
Sandaruwan